pyproject.toml
[project]
name = "pydantic-ai-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.2.0",
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "langchain-google-genai>=2.1.2",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.3.31",
    "logfire>=3.14.0",
    "mcp==1.6.0",
    "openai-agents>=0.0.12",
    "pydantic-ai-slim[mcp]>=0.1.3",
    "pydantic-evals[logfire]>=0.1.3",
    "python-dotenv>=1.1.0",
]
